+++ lib/kline_chart.dart
@@ -1,5 +1,6 @@
 library flutter_kline;
 
 export 'src/kline_view.dart';
+export 'src/kline_chart_style.dart';
 export 'src/kline_controller.dart';
 export 'src/kline_data.dart';